test: support tests with different extensions than the lang name
previously, we only ran tests for when the treesitter language name is
the same as the extension name, but this is not the case for many
languages. in this commit, we add support for these excluded tests. For
example, test.sol was excluded because the test searches for
test.solidity which does not exist. Now, we get all files in test/lang/*
and group them by filetype, get the filetype for each treesitter
language, and run the tests that match the filetype.